Added two additional features to the Fixation Filtration - the ability to a pre-made query from a file, and the ability to save filtration settings as a file.
